+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Dropped support for Dart 1. Going forward only Dart 2 will be supported.
   * If you experience blocking issues and are still on the Dart 1 sdk, we will
     consider bug fixes on a per-case basis based on severity and impact.
+  * Drop support for `dartium` and `content-shell` platforms since those are
+    removed from the Dart 2 SDK.
 * Fixed an issue `--precompiled` node tests in subdirectories.
 * Fixed some dart2 issues with node test bootstrapping code so that dartdevc
   tests can run.